move extractors into separate module, isolate API layer
