+++ avs-api/src/testFixtures/java/org/example/age/test/avs/service/StubAvsServiceModule.java
@@ -4,8 +4,8 @@ import dagger.Binds;
 import dagger.Module;
 import org.example.age.avs.api.AvsApi;
 import org.example.age.avs.api.AvsApiModule;
-import org.example.age.common.service.data.UserAgentAuthMatchDataExtractorModule;
-import org.example.age.test.common.service.data.TestAccountIdExtractorModule;
+import org.example.age.common.api.extractor.builtin.UserAgentAuthMatchDataExtractorModule;
+import org.example.age.common.api.extractor.test.TestAccountIdExtractorModule;
 
 /** Dagger module that binds dependencies for <code>@Named("api") HttpHandler</code>. */
 @Module(
