--- common-data/src/main/java/org/example/age/common/service/data/UserAgentAuthMatchDataExtractorModule.java
@@ -1,16 +1,0 @@
-package org.example.age.common.service.data;
-
-import dagger.Binds;
-import dagger.Module;
-import org.example.age.common.api.data.AuthMatchDataExtractor;
-
-/**
- * Dagger module that publishes a binding for {@link AuthMatchDataExtractor},
- * which matches the {@code User-Agent} header.
- */
-@Module
-public interface UserAgentAuthMatchDataExtractorModule {
-
-    @Binds
-    AuthMatchDataExtractor bindAuthMatchDataExtractor(UserAgentAuthMatchDataExtractor impl);
-}
