+++ common-extractor-builtin/src/main/java/org/example/age/common/api/extractor/builtin/DisabledAuthMatchDataExtractor.java
@@ -1,0 +1,22 @@
+package org.example.age.common.api.extractor.builtin;
+
+import io.undertow.server.HttpServerExchange;
+import javax.inject.Inject;
+import javax.inject.Singleton;
+import org.example.age.api.HttpOptional;
+import org.example.age.common.api.data.AuthMatchData;
+import org.example.age.common.api.extractor.AuthMatchDataExtractor;
+
+/** {@link AuthMatchDataExtractor} that extracts {@link DisabledAuthMatchData}. */
+@Singleton
+final class DisabledAuthMatchDataExtractor implements AuthMatchDataExtractor {
+
+    @Inject
+    public DisabledAuthMatchDataExtractor() {}
+
+    @Override
+    public HttpOptional<AuthMatchData> tryExtract(HttpServerExchange exchange) {
+        AuthMatchData authData = DisabledAuthMatchData.of();
+        return HttpOptional.of(authData);
+    }
+}
