+++ common-extractor-builtin/src/main/java/org/example/age/common/api/extractor/builtin/UserAgentAuthMatchDataExtractor.java
@@ -1,0 +1,25 @@
+package org.example.age.common.api.extractor.builtin;
+
+import io.undertow.server.HttpServerExchange;
+import io.undertow.util.Headers;
+import javax.inject.Inject;
+import javax.inject.Singleton;
+import org.example.age.api.HttpOptional;
+import org.example.age.common.api.data.AuthMatchData;
+import org.example.age.common.api.extractor.AuthMatchDataExtractor;
+
+/** {@link AuthMatchDataExtractor} that extracts {@link UserAgentAuthMatchData}. */
+@Singleton
+final class UserAgentAuthMatchDataExtractor implements AuthMatchDataExtractor {
+
+    @Inject
+    public UserAgentAuthMatchDataExtractor() {}
+
+    @Override
+    public HttpOptional<AuthMatchData> tryExtract(HttpServerExchange exchange) {
+        String userAgent = exchange.getRequestHeaders().getFirst(Headers.USER_AGENT);
+        userAgent = (userAgent != null) ? userAgent : "";
+        AuthMatchData authData = UserAgentAuthMatchData.of(userAgent);
+        return HttpOptional.of(authData);
+    }
+}
