+++ common-extractor-builtin/src/testFixtures/java/org/example/age/common/api/extractor/test/TestAccountIdExtractor.java
@@ -1,0 +1,24 @@
+package org.example.age.common.api.extractor.test;
+
+import io.undertow.server.HttpServerExchange;
+import io.undertow.util.StatusCodes;
+import java.util.Optional;
+import javax.inject.Inject;
+import javax.inject.Singleton;
+import org.example.age.api.HttpOptional;
+import org.example.age.common.api.extractor.AccountIdExtractor;
+
+/** Extracts an account ID from the custom {@code Account-Id} header, or sends a 401 error. */
+@Singleton
+final class TestAccountIdExtractor implements AccountIdExtractor {
+
+    @Inject
+    public TestAccountIdExtractor() {}
+
+    @Override
+    public HttpOptional<String> tryExtract(HttpServerExchange exchange) {
+        Optional<String> maybeAccountId =
+                Optional.ofNullable(exchange.getRequestHeaders().getFirst("Account-Id"));
+        return HttpOptional.fromOptional(maybeAccountId, StatusCodes.UNAUTHORIZED);
+    }
+}
