+++ common/src/test/java/org/example/age/common/avs/verification/VerificationManagerTest.java
@@ -17,6 +17,7 @@ import javax.inject.Named;
 import javax.inject.Singleton;
 import org.assertj.core.data.Offset;
 import org.example.age.api.HttpOptional;
+import org.example.age.common.api.extractor.builtin.DisabledAuthMatchDataExtractorModule;
 import org.example.age.common.avs.config.RegisteredSiteConfig;
 import org.example.age.common.avs.store.InMemoryRegisteredSiteConfigStoreModule;
 import org.example.age.common.avs.store.InMemoryVerifiedUserStoreModule;
@@ -25,7 +26,6 @@ import org.example.age.common.avs.store.VerifiedUserStore;
 import org.example.age.common.avs.verification.internal.Verification;
 import org.example.age.common.avs.verification.internal.VerificationManager;
 import org.example.age.common.avs.verification.internal.VerificationManagerModule;
-import org.example.age.common.service.data.DisabledAuthMatchDataExtractorModule;
 import org.example.age.common.service.data.SiteLocation;
 import org.example.age.common.service.store.InMemoryPendingStoreFactoryModule;
 import org.example.age.data.certificate.VerificationSession;
