+++ infra-api/build.gradle.kts
@@ -1,7 +1,6 @@
 plugins {
     id("org.example.age.java-conventions")
     `java-library`
-    `java-test-fixtures`
 }
 
 dependencies {
@@ -9,6 +8,7 @@ dependencies {
     annotationProcessor("com.google.dagger:dagger-compiler")
 
     api(project(":api"))
+    api("com.fasterxml.jackson.core:jackson-core")
     api("com.fasterxml.jackson.core:jackson-databind")
     api("com.google.dagger:dagger")
     api("io.undertow:undertow-core")
