+++ infra-api/src/main/java/org/example/age/infra/api/RequestParser.java
@@ -29,8 +29,8 @@ public final class RequestParser {
      * <p>Sends a 400 error if the body cannot be deserialized,
      * or a 500 error if the callback throws an uncaught exception.</p>
      */
-    public <B> void readBody(TypeReference<B> bodyTypeRef, RequestJsonCallback<B> callback) {
-        Receiver.FullBytesCallback adaptedCallback = new AdaptedRequestJsonCallback<>(bodyTypeRef, callback);
+    public <V> void readBody(TypeReference<V> valueTypeRef, RequestJsonCallback<V> callback) {
+        Receiver.FullBytesCallback adaptedCallback = new AdaptedRequestJsonCallback<>(valueTypeRef, callback);
         exchange.getRequestReceiver().receiveFullBytes(adaptedCallback);
     }
 
@@ -39,7 +39,7 @@ public final class RequestParser {
      *
      * <p>Returns a 400 error if the parameter is missing or cannot be deserialized.</p>
      */
-    public <P> HttpOptional<P> tryGetQueryParameter(String name, TypeReference<P> paramTypeRef) {
+    public <V> HttpOptional<V> tryGetQueryParameter(String name, TypeReference<V> valueTypeRef) {
         HttpOptional<String> maybeRawValue = tryGetQueryParameter(name);
         if (maybeRawValue.isEmpty()) {
             return HttpOptional.empty(maybeRawValue.statusCode());
@@ -47,7 +47,7 @@ public final class RequestParser {
 
         String rawValue = maybeRawValue.get();
         byte[] json = serializer.serialize(rawValue);
-        return serializer.tryDeserialize(json, paramTypeRef, StatusCodes.BAD_REQUEST);
+        return serializer.tryDeserialize(json, valueTypeRef, StatusCodes.BAD_REQUEST);
     }
 
     /**
@@ -70,28 +70,28 @@ public final class RequestParser {
     }
 
     /** Adapts a {@link RequestJsonCallback} to a {@link Receiver.FullBytesCallback}. */
-    private final class AdaptedRequestJsonCallback<B> implements Receiver.FullBytesCallback {
+    private final class AdaptedRequestJsonCallback<V> implements Receiver.FullBytesCallback {
 
-        private final TypeReference<B> bodyTypeRef;
-        private final RequestJsonCallback<B> callback;
+        private final TypeReference<V> valueTypeRef;
+        private final RequestJsonCallback<V> callback;
 
-        public AdaptedRequestJsonCallback(TypeReference<B> bodyTypeRef, RequestJsonCallback<B> callback) {
-            this.bodyTypeRef = bodyTypeRef;
+        public AdaptedRequestJsonCallback(TypeReference<V> valueTypeRef, RequestJsonCallback<V> callback) {
+            this.valueTypeRef = valueTypeRef;
             this.callback = callback;
         }
 
         @Override
-        public void handle(HttpServerExchange exchange, byte[] rawBody) {
-            HttpOptional<B> maybeBody = serializer.tryDeserialize(rawBody, bodyTypeRef, StatusCodes.BAD_REQUEST);
-            if (maybeBody.isEmpty()) {
-                sendErrorCode(maybeBody.statusCode());
+        public void handle(HttpServerExchange exchange, byte[] rawValue) {
+            HttpOptional<V> maybeValue = serializer.tryDeserialize(rawValue, valueTypeRef, StatusCodes.BAD_REQUEST);
+            if (maybeValue.isEmpty()) {
+                sendErrorCode(maybeValue.statusCode());
                 return;
             }
-            B body = maybeBody.get();
+            V value = maybeValue.get();
 
             // FullBytesCallback does not throw an exception, so we must handle exceptions here.
             try {
-                callback.handleRequest(exchange, RequestParser.this, body);
+                callback.handleRequest(exchange, RequestParser.this, value);
             } catch (Exception e) {
                 sendErrorCode(StatusCodes.INTERNAL_SERVER_ERROR);
             }
