+++ site-api/src/main/java/org/example/age/site/api/SiteEndpointHandler.java
@@ -12,9 +12,9 @@ import org.example.age.api.HttpOptional;
 import org.example.age.api.JsonSender;
 import org.example.age.api.JsonSerializer;
 import org.example.age.api.StatusCodeSender;
-import org.example.age.common.api.data.AccountIdExtractor;
 import org.example.age.common.api.data.AuthMatchData;
-import org.example.age.common.api.data.AuthMatchDataExtractor;
+import org.example.age.common.api.extractor.AccountIdExtractor;
+import org.example.age.common.api.extractor.AuthMatchDataExtractor;
 import org.example.age.data.certificate.SignedAgeCertificate;
 import org.example.age.data.certificate.VerificationSession;
 import org.example.age.infra.api.ExchangeDispatcher;
