+++ site-api/src/testFixtures/java/org/example/age/test/site/service/StubSiteServiceModule.java
@@ -2,10 +2,10 @@ package org.example.age.test.site.service;
 
 import dagger.Binds;
 import dagger.Module;
-import org.example.age.common.service.data.UserAgentAuthMatchDataExtractorModule;
+import org.example.age.common.api.extractor.builtin.UserAgentAuthMatchDataExtractorModule;
+import org.example.age.common.api.extractor.test.TestAccountIdExtractorModule;
 import org.example.age.site.api.SiteApi;
 import org.example.age.site.api.SiteApiModule;
-import org.example.age.test.common.service.data.TestAccountIdExtractorModule;
 
 /** Dagger module that binds dependencies for <code>@Named("api") HttpHandler</code>. */
 @Module(
