+++ site-service/src/test/java/org/example/age/site/service/SiteServiceTest.java
@@ -15,8 +15,9 @@ import java.util.Map;
 import javax.inject.Named;
 import javax.inject.Singleton;
 import org.example.age.api.HttpOptional;
+import org.example.age.common.api.extractor.builtin.DisabledAuthMatchDataExtractorModule;
+import org.example.age.common.api.extractor.test.TestAccountIdExtractorModule;
 import org.example.age.common.service.data.AvsLocation;
-import org.example.age.common.service.data.DisabledAuthMatchDataExtractorModule;
 import org.example.age.common.service.store.InMemoryPendingStoreFactoryModule;
 import org.example.age.data.certificate.VerificationSession;
 import org.example.age.data.crypto.SecureId;
@@ -24,7 +25,6 @@ import org.example.age.site.service.config.SiteConfig;
 import org.example.age.site.service.store.InMemoryVerificationStoreModule;
 import org.example.age.test.avs.service.FakeAvsServiceModule;
 import org.example.age.test.common.service.crypto.TestSigningKeyModule;
-import org.example.age.test.common.service.data.TestAccountIdExtractorModule;
 import org.example.age.test.common.service.data.TestAvsLocationModule;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.server.TestServer;
