+++ site-service/src/test/java/org/example/age/site/service/verification/internal/VerificationManagerTest.java
@@ -12,8 +12,8 @@ import javax.inject.Named;
 import javax.inject.Singleton;
 import org.assertj.core.data.Offset;
 import org.example.age.common.api.data.AuthMatchData;
+import org.example.age.common.api.extractor.builtin.UserAgentAuthMatchData;
 import org.example.age.common.service.crypto.internal.AuthMatchDataEncryptor;
-import org.example.age.common.service.data.UserAgentAuthMatchData;
 import org.example.age.common.service.data.internal.DataMapperModule;
 import org.example.age.common.service.store.InMemoryPendingStoreFactoryModule;
 import org.example.age.data.certificate.AgeCertificate;
