+++ site-service/src/testFixtures/java/org/example/age/test/avs/service/FakeAvsServiceModule.java
@@ -4,11 +4,11 @@ import dagger.Binds;
 import dagger.Module;
 import org.example.age.avs.api.AvsApi;
 import org.example.age.avs.api.AvsApiModule;
-import org.example.age.common.service.data.DisabledAuthMatchDataExtractorModule;
+import org.example.age.common.api.extractor.builtin.DisabledAuthMatchDataExtractorModule;
+import org.example.age.common.api.extractor.test.TestAccountIdExtractorModule;
 import org.example.age.common.service.data.internal.DataMapperModule;
 import org.example.age.infra.service.client.RequestDispatcherModule;
 import org.example.age.test.avs.service.verification.internal.FakeAvsVerificationFactoryModule;
-import org.example.age.test.common.service.data.TestAccountIdExtractorModule;
 import org.example.age.test.common.service.data.TestSiteLocationModule;
 import org.example.age.testing.server.TestServer;
