--- testing-api/build.gradle.kts
@@ -1,10 +1,0 @@
-plugins {
-    id("org.example.age.java-conventions")
-    `java-test-fixtures`
-}
-
-dependencies {
-    // test fixtures
-    testFixturesApi("io.undertow:undertow-core")
-    testFixturesApi("org.mockito:mockito-core")
-}
